requirements.txt
typer
moviepy
openai
azure.cognitiveservices.speech
py_linq
python-dotenv